Extract nodes and users into separate mods
