+++ test-utils/testlib/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod alphanet_utils;
-pub mod node_user;
-pub mod test_locks;
+pub mod node;
+pub mod user;
+pub mod test_helpers;
