+++ test-utils/testlib/src/test_helpers.rs
@@ -1,0 +1,64 @@
+use lazy_static::lazy_static;
+use std::sync::{Mutex, Arc, RwLock};
+use std::process::Output;
+use std::thread;
+use std::time::Duration;
+use crate::node::Node;
+
+lazy_static! {
+    static ref HEAVY_TESTS_LOCK: Mutex<()> = Mutex::new(());
+}
+
+pub fn heavy_test<F>(f: F)
+where
+    F: FnOnce() -> (),
+{
+    let _guard = HEAVY_TESTS_LOCK.lock();
+    f();
+}
+
+pub fn check_result(output: Output) -> Result<String, String> {
+    let mut result = String::from_utf8_lossy(output.stdout.as_slice());
+    if !output.status.success() {
+        if result.is_empty() {
+            result = String::from_utf8_lossy(output.stderr.as_slice());
+        }
+        return Err(result.to_owned().to_string());
+    }
+    Ok(result.to_owned().to_string())
+}
+
+pub fn wait<F>(f: F, check_interval_ms: u64, max_wait_ms: u64)
+    where
+        F: Fn() -> bool,
+{
+    let mut ms_slept = 0;
+    while !f() {
+        thread::sleep(Duration::from_millis(check_interval_ms));
+        ms_slept += check_interval_ms;
+        if ms_slept > max_wait_ms {
+            println!("BBBB Slept {}; max_wait_ms {}", ms_slept, max_wait_ms);
+            panic!("Timed out waiting for the condition");
+        }
+    }
+}
+
+/// TODO it makes sense to have three types of wait checks:
+/// Wait until sufficient number of nodes is caught up (> 2/3). This can be checked by looking at the block indices and verifying that the blocks are produced;
+/// Wait until a certain node is caught up and participating in a consensus. Check first-layer BLS signatures;
+/// Wait until all nodes are more-or-less caught up. Check that the max_block_index - min_block_index < threshold;
+///
+pub fn wait_for_catchup(nodes: &Vec<Arc<RwLock<dyn Node>>>) {
+    wait(
+        || {
+            let tips: Vec<_> = nodes
+                .iter()
+                .filter(|node| node.read().unwrap().is_running())
+                .map(|node| node.read().unwrap().user().get_best_block_index())
+                .collect();
+            tips.iter().min() == tips.iter().max()
+        },
+        1000,
+        10000,
+    );
+}
