--- test-utils/testlib/src/test_locks.rs
@@ -1,14 +1,0 @@
-use lazy_static::lazy_static;
-use std::sync::Mutex;
-
-lazy_static! {
-    static ref HEAVY_TESTS_LOCK: Mutex<()> = Mutex::new(());
-}
-
-pub fn heavy_test<F>(f: F)
-where
-    F: FnOnce() -> (),
-{
-    let _guard = HEAVY_TESTS_LOCK.lock();
-    f();
-}
